FIX: Change packet downloader to cron-based rather than event based
The event based approach was too simplistic and didn't account for
multiple files arriving. For now, people want their data sooner, so
run the data retrieval every 6-hours and create packet files as
needed during those periods only if there was data received on the
ground. This could lead to more version bumps than necessary, but
tries to strike a reasonable position of not having to wait too
long to get packets either.